fix : ItemDbId type & Managing Id
Item Db Id를 long타입으로 수정,
id를 어플리케이션에서 관리하도록 수정

+++ SuperServer/Game/Room/GameRoom_Item.cs
@@ -21,7 +21,7 @@ namespace SuperServer.Game.Room
 
             dropItem.Pickup(picker);
         }
-        public void HandleUseItem(Hero hero, int itemDbId)
+        public void HandleUseItem(Hero hero, long itemDbId)
         {
             Item item = hero.Inventory.GetItemByDbId(itemDbId);
 
@@ -33,7 +33,7 @@ namespace SuperServer.Game.Room
                 return;
         }
 
-        public void HandleEquipItem(Hero hero, int itemDbId)
+        public void HandleEquipItem(Hero hero, long itemDbId)
         {
             Item item = hero.Inventory.GetItemByDbId(itemDbId);
 
@@ -45,7 +45,7 @@ namespace SuperServer.Game.Room
             equipItem.EquipItem(hero);
         }
 
-        public void HandleUnEquipItem(Hero hero, int itemDbId)
+        public void HandleUnEquipItem(Hero hero, long itemDbId)
         {
             Item item = hero.Inventory.GetItemByDbId(itemDbId);
